requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
sentence-transformers==2.6.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
ollama==0.1.7
requests==2.31.0

pypdf2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.3

huggingface-hub==0.20.3
torch==2.1.2
transformers==4.37.2

chromadb>=0.4.0
scikit-learn==1.4.0